Move wal_stream_connection_config to utils.
It will be used by safekeeper as well.
